frontend/RAG ChatBot/requirements.txt
flask==3.0.0
pdfplumber==0.11.0
openai==1.54.0
faiss-cpu==1.8.0
numpy==1.26.0
pandas==2.1.0
python-dotenv==1.0.0
